% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sampled_multi.R
\name{get_sampled_multi}
\alias{get_sampled_multi}
\title{Title}
\usage{
get_sampled_multi(
  TAG,
  start_time = "*-7d",
  end_time = "*",
  freq = "1m",
  isWebID = FALSE
)
}
\arguments{
\item{TAG}{a list of PI tags}

\item{start_time}{a date-time object (lubridate format)}

\item{end_time}{a date-time object (lubridate format)}

\item{freq}{return resolution ("1m", "1h", "1d", etc... )}

\item{isWebID}{boolean, TRUE if TAG is a WebID}
}
\value{
data.frame
}
\description{
Title
}
\examples{
tags <- c("wwl:east:bra10477B_REALTMMETFLO")

get_sampled_multi(tags, mdy("12/16/2021"), mdy("1/31/2022"))
}
