% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_webid.R
\name{get_webid}
\alias{get_webid}
\title{Get WebID}
\usage{
get_webid(
  pi_tag,
  pi_server = "masterpi",
  servername = "https://MASTERPIAPP.corp.jea.com/piwebapi/",
  is_data_server = TRUE
)
}
\arguments{
\item{is_data_server}{}
}
\description{
Primary use case is to determine if a tag exists
}
\examples{
  get_webid("wwl:stjohns:wam213_stpwp")
}
